# src/main/resources/api/openapi.yaml
openapi: 3.0.4
info:
  title: Product Catalog API
  description: API REST para gestión de catálogo de productos de e-commerce
  version: 1.0.0
  contact:
    name: ITQ distributed and cloud systems
    email: ivonne.al@queretaro.tecnm.mx
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo
  - url: https://api.ecommerce.com/api/v1
    description: Servidor de producción

tags:
  - name: products
    description: Operaciones relacionadas con catálogo de productos

paths:
  /products:
    post:
      summary: Crear nuevo producto
      description: Crea un nuevo producto en el catálogo
      operationId: createProduct
      tags:
        - products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInvalidInput:
                  $ref: '#/components/examples/ApiErrorInvalidInput'
        '409':
          description: Conflicto - producto ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorConflict:
                  $ref: '#/components/examples/ApiErrorConflict'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInternal:
                  $ref: '#/components/examples/ApiErrorInternal'

    get:
      summary: Buscar productos
      description: Busca productos existentes basado en criterios
      operationId: searchProducts
      tags:
        - products
      parameters:
        - name: category
          in: query
          description: Categoría del producto
          required: false
          schema:
            type: string
            enum: ["Electronics", "Clothing", "Books", "Home", "Sports"]
          example: "Electronics"
        - name: minPrice
          in: query
          description: Precio mínimo del producto
          required: false
          schema:
            type: number
            format: double
            minimum: 0
          example: 100.00
        - name: maxPrice
          in: query
          description: Precio máximo del producto
          required: false
          schema:
            type: number
            format: double
            minimum: 0
          example: 2000.00
        - name: inStock
          in: query
          description: Solo productos en stock
          required: false
          schema:
            type: boolean
          example: true
        - name: page
          in: query
          description: Número de página (por defecto 1)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Elementos por página (por defecto 20, máximo 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
      responses:
        '200':
          description: Lista de productos encontrados (puede estar vacía)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInvalidInput:
                  $ref: '#/components/examples/ApiErrorInvalidInput'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInternal:
                  $ref: '#/components/examples/ApiErrorInternal'

  /products/{productId}:
    get:
      summary: Obtener producto por ID
      description: Obtiene los detalles de un producto específico
      operationId: getProductById
      tags:
        - products
      parameters:
        - name: productId
          in: path
          description: ID del producto a obtener
          required: true
          schema:
            type: string
            pattern: '^prod-[0-9]+$'
          example: "prod-12345"
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorNotFound:
                  $ref: '#/components/examples/ApiErrorNotFound'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInternal:
                  $ref: '#/components/examples/ApiErrorInternal'

    put:
      summary: Actualizar producto completo
      description: Actualiza todos los datos de un producto existente
      operationId: updateProduct
      tags:
        - products
      parameters:
        - name: productId
          in: path
          description: ID del producto a actualizar
          required: true
          schema:
            type: string
            pattern: '^prod-[0-9]+$'
          example: "prod-12345"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInvalidInput:
                  $ref: '#/components/examples/ApiErrorInvalidInput'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorNotFound:
                  $ref: '#/components/examples/ApiErrorNotFound'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInternal:
                  $ref: '#/components/examples/ApiErrorInternal'

    patch:
      summary: Actualizar producto parcial
      description: Actualiza solo algunos campos de un producto existente
      operationId: patchProduct
      tags:
        - products
      parameters:
        - name: productId
          in: path
          description: ID del producto a actualizar
          required: true
          schema:
            type: string
            pattern: '^prod-[0-9]+$'
          example: "prod-12345"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductPatch'
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInvalidInput:
                  $ref: '#/components/examples/ApiErrorInvalidInput'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorNotFound:
                  $ref: '#/components/examples/ApiErrorNotFound'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInternal:
                  $ref: '#/components/examples/ApiErrorInternal'

    delete:
      summary: Eliminar producto
      description: Elimina un producto del catálogo
      operationId: deleteProduct
      tags:
        - products
      parameters:
        - name: productId
          in: path
          description: ID del producto a eliminar
          required: true
          schema:
            type: string
            pattern: '^prod-[0-9]+$'
          example: "prod-12345"
      responses:
        '200':
          description: Producto eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteConfirmation'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorNotFound:
                  $ref: '#/components/examples/ApiErrorNotFound'
        '409':
          description: No se puede eliminar - producto tiene pedidos activos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorProductHasOrders:
                  $ref: '#/components/examples/ApiErrorProductHasOrders'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                apiErrorInternal:
                  $ref: '#/components/examples/ApiErrorInternal'

components:
  schemas:
    Category:
      type: string
      enum: [ELECTRONICS, CLOTHING, BOOKS, HOME, SPORTS]
      description: "Categoría del producto"
      example: "Electronics"

    Product:
      type: object
      required:
        - name
        - price  
        - category
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: "Nombre del producto"
          example: "Laptop Gaming"
        description:
          type: string
          maxLength: 1000
          description: "Descripción detallada del producto"
          example: "Laptop de alto rendimiento para gaming"
        price:
          type: number
          format: double
          minimum: 0
          description: "Precio del producto en USD"
          example: 1299.99
        category:
          $ref: '#/components/schemas/Category'
        inStock:
          type: boolean
          description: "Indica si el producto está disponible"
          example: true
          default: true

    ProductResponse:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          required:
            - id
            - sku
            - createdAt
            - inStock
          properties:
            id:
              type: string 
              pattern: '^prod-[0-9]+$'
              description: "ID único del producto"
              example: "prod-12345"
            sku:
              type: string
              description: "Código SKU del producto"
              example: "LAP-GAM-001"
            createdAt:
              type: string
              format: date-time
              description: "Fecha y hora de creación"
              example: "2024-01-15T10:30:00Z"
            updatedAt:
              type: string
              format: date-time
              description: "Fecha y hora de última actualización"
              example: "2024-01-15T10:30:00Z"

    ProductPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        price:
          type: number
          format: double
          minimum: 0
        category:
          type: string
          enum: ["Electronics", "Clothing", "Books", "Home", "Sports"]
        inStock:
          type: boolean

    ProductList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'

    Pagination:
      type: object
      required:
        - total
        - page
        - limit
        - totalPages
      properties:
        total:
          type: integer
          minimum: 0
          description: "Total de productos"
          example: 150
        page:
          type: integer
          minimum: 1
          description: "Página actual"
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: "Elementos por página"
          example: 20
        totalPages:
          type: integer
          minimum: 0
          description: "Total de páginas"
          example: 8

    DeleteConfirmation:
      type: object
      required:
        - id
        - message
        - deletedAt
      properties:
        id:
          type: string
          description: "ID del producto eliminado"
          example: "prod-12345"
        message:
          type: string
          description: "Mensaje de confirmación"
          example: "Producto eliminado exitosamente"
        deletedAt:
          type: string
          format: date-time
          description: "Fecha y hora de eliminación"
          example: "2024-01-15T15:45:00Z"

    ApiError:
      type: object
      description: "Estructura estándar para errores de la API"
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: |
            Código de error.
            Valores comunes en esta API:
            - INVALID_INPUT (400)
            - PRODUCT_NOT_FOUND (404)
            - PRODUCT_EXISTS (409 - conflicto en creación)
            - PRODUCT_HAS_ORDERS (409 - conflicto en eliminación)
            - INTERNAL_ERROR (500)
          enum:
            - INVALID_INPUT
            - PRODUCT_NOT_FOUND
            - PRODUCT_EXISTS
            - PRODUCT_HAS_ORDERS
            - INTERNAL_ERROR
          example: "PRODUCT_NOT_FOUND"
        message:
          type: string
          description: "Mensaje descriptivo del error"
          example: "El producto especificado no existe"
        details:
          type: object
          description: "Detalles adicionales del error"
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: "Fecha y hora del error"
          example: "2024-01-15T10:30:00Z"
        path:
          type: string
          description: "Ruta donde ocurrió el error"
          example: "/api/v1/products/prod-12345"
  examples:
    ApiErrorInvalidInput:
      summary: Datos de entrada inválidos (400)
      value:
        code: "INVALID_INPUT"
        message: "Los datos enviados no cumplen con el esquema."
        timestamp: "2024-01-15T10:30:00Z"
        path: "/api/v1/products"
        details:
          field: "name"
          reason: "required"
    ApiErrorNotFound:
      summary: Producto no encontrado (404)
      value:
        code: "PRODUCT_NOT_FOUND"
        message: "El producto especificado no existe"
        timestamp: "2024-01-15T10:30:00Z"
        path: "/api/v1/products/prod-12345"
        details:
          productId: "prod-12345"
          searchedBy: "id"
    ApiErrorConflict:
      summary: Conflicto - producto ya existe (409)
      value:
        code: "PRODUCT_EXISTS"
        message: "El producto ya existe."
        timestamp: "2024-01-15T10:30:00Z"
        path: "/api/v1/products"
        details:
          existingProductId: "prod-12345"
          conflictReason: "duplicate_name"
    ApiErrorProductHasOrders:
      summary: No se puede eliminar - producto tiene pedidos activos (409)
      value:
        code: "PRODUCT_HAS_ORDERS"
        message: "No se puede eliminar - producto tiene pedidos activos"
        timestamp: "2024-01-15T10:30:00Z"
        path: "/api/v1/products/prod-12345"
        details:
          activeOrdersCount: 3
          orderIds: ["ord-001", "ord-002", "ord-003"]
    ApiErrorInternal:
      summary: Error interno del servidor (500)
      value:
        code: "INTERNAL_ERROR"
        message: "Ocurrió un error inesperado."
        timestamp: "2024-01-15T10:30:00Z"
        path: "/api/v1/products"
        details:
          errorId: "err-2024-001"
          traceId: "trace-abc123"